fix: postman 테스트 중 -> 에러 발생한 api 수정 중.

+++ src/main/java/warehouseLocation/domain/controller/ProductController.java
@@ -31,7 +31,7 @@ public class ProductController {
 
     //  2.2 (GET) /manage/product/{product_id} : 상품 정보
     @GetMapping("/product/{productId}")
-    public ProductResDto.ProductInfo productInfo(@PathVariable Long productId) throws Exception {
+    public ProductResDto.ProductInfo productInfo(@PathVariable Integer productId) throws Exception {
         return this.productService.productInfo(productId);
     }
 
@@ -46,7 +46,7 @@ public class ProductController {
 
     //  //3.1(DELETE) /manage/product/{productId} : 해당 상품 삭제 (완전 삭제 대신, 업데이트로 진행)
     @DeleteMapping("/product/{productId}")
-    public ResponseEntity<ProductResDto.Message> softDeleteProduct(@PathVariable Long productId) {
+    public ResponseEntity<ProductResDto.Message> softDeleteProduct(@PathVariable Integer productId) {
         return this.productService.softDeleteProduct(productId);
     }
